+++ lib/shrine/plugins/activerecord.rb
@@ -10,7 +10,7 @@ class Shrine
     module Activerecord
       def self.load_dependencies(uploader, **)
         uploader.plugin :model
-        uploader.plugin :atomic_helpers
+        uploader.plugin :_persistence, plugin: self
       end
 
       def self.configure(uploader, **opts)
@@ -27,12 +27,12 @@ class Shrine
           name = @name
 
           if shrine_class.opts[:activerecord][:validations]
+            # add validation plugin integration
             model.validate do
-              # validation plugin integration
-              if send(:"#{name}_attacher").respond_to?(:errors)
-                send(:"#{name}_attacher").errors.each do |message|
-                  errors.add(name, *message)
-                end
+              next unless send(:"#{name}_attacher").respond_to?(:errors)
+
+              send(:"#{name}_attacher").errors.each do |message|
+                errors.add(name, *message)
               end
             end
           end
@@ -48,7 +48,7 @@ class Shrine
               model.after_commit on: action do
                 if send(:"#{name}_attacher").changed?
                   send(:"#{name}_attacher").finalize
-                  send(:"#{name}_attacher").activerecord_persist
+                  send(:"#{name}_attacher").persist
                 end
               end
             end
@@ -68,79 +68,23 @@ class Shrine
       end
 
       module AttacherMethods
-        # Promotes cached file to permanent storage in an atomic way. It's
-        # intended to be called from a background job.
+        # The _persistence plugin defines the following methods:
         #
-        #     attacher.assign(file)
-        #     attacher.cached? #=> true
-        #
-        #     # ... in background job ...
-        #
-        #     attacher.atomic_promote
-        #     attacher.stored? #=> true
-        #
-        # It accepts `:reload` and `:persist` strategies:
-        #
-        #     attacher.atomic_promote(reload: :lock)    # uses database locking (default)
-        #     attacher.atomic_promote(reload: :fetch)   # reloads with no locking
-        #     attacher.atomic_promote(reload: ->(&b){}) # custom reloader
-        #     attacher.atomic_promote(reload: false)    # skips reloading
-        #
-        #     attacher.atomic_promote(persist: :save) # persists stored file (default)
-        #     attacher.atomic_promote(persist: ->{})  # custom persister
-        #     attacher.atomic_promote(persist: false) # skips persistence
-        def activerecord_atomic_promote(**options, &block)
-          abstract_atomic_promote(activerecord_strategies(**options), &block)
-        end
-        alias atomic_promote activerecord_atomic_promote
-
-        # Persist the the record only if the attachment hasn't changed.
-        # Optionally yields reloaded attacher to the block before persisting.
-        # It's intended to be called from a background job.
-        #
-        #     # ... in background job ...
-        #
-        #     attacher.file.metadata["foo"] = "bar"
-        #     attacher.write
-        #
-        #     attacher.atomic_persist
-        def activerecord_atomic_persist(*args, **options, &block)
-          abstract_atomic_persist(*args, activerecord_strategies(**options), &block)
-        end
-        alias atomic_persist activerecord_atomic_persist
-
-        # Called in the `after_commit` callback after finalization.
-        def activerecord_persist
-          activerecord_save
-        end
-        alias persist activerecord_persist
-
+        #   * #persist (calls #activerecord_persist and #activerecord?)
+        #   * #atomic_persist (calls #activerecord_lock, #activerecord_persist and #activerecord?)
+        #   * #atomic_promote (calls #activerecord_lock, #activerecord_persist and #activerecord?)
         private
 
-        # Resolves strategies for atomic promotion and persistence.
-        def activerecord_strategies(reload: :lock, persist: :save, **options)
-          reload  = method(:"activerecord_#{reload}")  if reload.is_a?(Symbol)
-          persist = method(:"activerecord_#{persist}") if persist.is_a?(Symbol)
-
-          { reload: reload, persist: persist, **options }
-        end
-
-        # Implements the "fetch" reload strategy for #atomic_promote and
-        # #atomic_persist.
-        def activerecord_fetch
-          yield record.clone.reload
-        end
-
-        # Implements the "lock" reload strategy for #atomic_promote and
-        # #atomic_persist.
-        def activerecord_lock
-          record.transaction { yield record.clone.reload(lock: true) }
-        end
-
-        # Implements the "save" persist strategy for #atomic_promote and
-        # #atomic_persist.
-        def activerecord_save
+        # Saves changes to the model instance, skipping validations. Used by
+        # the _persistence plugin.
+        def activerecord_persist
           record.save(validate: false)
+        end
+
+        # Locks the database row and yields the reloaded record. Used by the
+        # _persistence plugin.
+        def activerecord_reload
+          record.transaction { yield record.clone.reload(lock: true) }
         end
 
         # ActiveRecord JSON column attribute needs to be assigned with a Hash.
@@ -150,10 +94,16 @@ class Shrine
 
         # Returns true if the data attribute represents a JSON or JSONB column.
         def activerecord_json_column?
-          return false unless record.is_a?(ActiveRecord::Base)
+          return false unless activerecord?
           return false unless column = record.class.columns_hash[attribute.to_s]
 
           [:json, :jsonb].include?(column.type)
+        end
+
+        # Returns whether the record is an ActiveRecord model. Used by the
+        # _persistence plugin.
+        def activerecord?
+          record.is_a?(::ActiveRecord::Base)
         end
       end
     end