+++ test/plugin/activerecord_test.rb
@@ -246,12 +246,12 @@ describe Shrine::Plugins::Activerecord do
   end
 
   describe "Attacher" do
-    describe "#activerecord_atomic_promote" do
+    describe "#atomic_promote" do
       it "promotes cached file to permanent storage" do
         @attacher.attach_cached(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_promote
+        @attacher.atomic_promote
 
         assert_equal :store, @attacher.file.storage_key
       end
@@ -260,7 +260,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach_cached(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_promote
+        @attacher.atomic_promote
 
         @attacher.reload
         assert_equal :store, @attacher.file.storage_key
@@ -274,7 +274,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach_cached(fakeio)
         @user.save
 
-        file = @attacher.activerecord_atomic_promote
+        file = @attacher.atomic_promote
 
         assert_equal @attacher.file, file
       end
@@ -283,7 +283,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach_cached(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_promote(location: "foo")
+        @attacher.atomic_promote(location: "foo")
 
         assert_equal "foo", @attacher.file.id
       end
@@ -293,7 +293,7 @@ describe Shrine::Plugins::Activerecord do
         @user.save
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_promote
+        @attacher.atomic_promote
 
         assert_equal "Janko", @user.name
         assert_equal "Janko", @user.reload.name
@@ -303,7 +303,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach_cached(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_promote { @user.name = "Janko" }
+        @attacher.atomic_promote { @user.name = "Janko" }
 
         assert_equal "Janko", @user.name
         assert_equal "Janko", @user.reload.name
@@ -316,7 +316,7 @@ describe Shrine::Plugins::Activerecord do
         @user.class.update_all(avatar_data: nil)
 
         assert_raises(Shrine::AttachmentChanged) do
-          @attacher.activerecord_atomic_promote { @block_called = true }
+          @attacher.atomic_promote { @block_called = true }
         end
 
         @user.reload
@@ -337,25 +337,12 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach_cached(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_promote
+        @attacher.atomic_promote
 
         @user.reload
         @attacher.reload
 
         assert_equal :store, @attacher.file.storage_key
-      end
-
-      it "accepts :fetch reload strategy" do
-        @attacher.attach_cached(fakeio)
-        @user.save
-
-        @user.name = "Janko"
-        @attacher.activerecord_atomic_promote(reload: :fetch)
-
-        @attacher.reload
-
-        assert_equal :store, @attacher.file.storage_key
-        assert_equal "Janko", @user.name
       end
 
       it "accepts custom reload strategy" do
@@ -365,7 +352,7 @@ describe Shrine::Plugins::Activerecord do
         @user.class.update_all(avatar_data: nil) # this change will not be detected
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_promote(reload: -> (&block) {
+        @attacher.atomic_promote(reload: -> (&block) {
           block.call @user.class.new(avatar_data: cached_file.to_json)
         })
 
@@ -383,7 +370,7 @@ describe Shrine::Plugins::Activerecord do
         @user.class.update_all(avatar_data: nil) # this change will not be detected
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_promote(reload: false)
+        @attacher.atomic_promote(reload: false)
 
         @user.reload
         @attacher.reload
@@ -396,7 +383,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach_cached(fakeio)
         @user.save(validate: false)
 
-        @attacher.activerecord_atomic_promote(persist: -> {
+        @attacher.atomic_promote(persist: -> {
           @user.name = "Janko"
           @user.save
         })
@@ -413,7 +400,7 @@ describe Shrine::Plugins::Activerecord do
         @user.save
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_promote(persist: false)
+        @attacher.atomic_promote(persist: false)
 
         assert_equal :store, @attacher.file.storage_key
         assert_equal "Janko", @user.name
@@ -425,26 +412,22 @@ describe Shrine::Plugins::Activerecord do
         assert_nil @user.name
       end
 
-      it "is aliased to #atomic_promote" do
-        @attacher.attach_cached(fakeio)
-        @user.save
+      it "raises NotImplementedError for non-ActiveRecord attacher" do
+        @attacher = @shrine::Attacher.new
 
-        @attacher.atomic_promote
-
-        @user.reload
-        @attacher.reload
-
-        assert_equal :store, @attacher.file.storage_key
+        assert_raises NotImplementedError do
+          @attacher.atomic_promote
+        end
       end
     end
 
-    describe "#activerecord_atomic_persist" do
+    describe "#atomic_persist" do
       it "persists the record" do
         file = @attacher.attach(fakeio)
         @user.save
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_persist
+        @attacher.atomic_persist
 
         assert_equal "Janko", @user.name
         assert_equal "Janko", @user.reload.name
@@ -455,7 +438,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_persist { @user.name = "Janko" }
+        @attacher.atomic_persist { @user.name = "Janko" }
 
         assert_equal "Janko", @user.name
         assert_equal "Janko", @user.reload.name
@@ -469,7 +452,7 @@ describe Shrine::Plugins::Activerecord do
 
         @user.name = "Janko"
         assert_raises(Shrine::AttachmentChanged) do
-          @attacher.activerecord_atomic_persist { @block_called = true }
+          @attacher.atomic_persist { @block_called = true }
         end
 
         @user.reload
@@ -492,22 +475,12 @@ describe Shrine::Plugins::Activerecord do
         @user.save
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_persist
+        @attacher.atomic_persist
 
         @user.reload
         @attacher.reload
 
         assert_equal "Janko", @user.name
-      end
-
-      it "accepts :fetch reload strategy" do
-        @attacher.attach(fakeio)
-        @user.save
-
-        @user.name = "Name"
-        @attacher.activerecord_atomic_persist(reload: :fetch)
-
-        assert_equal "Name", @user.name
       end
 
       it "accepts custom reload strategy" do
@@ -517,7 +490,7 @@ describe Shrine::Plugins::Activerecord do
         @user.class.update_all(avatar_data: nil) # this change will not be detected
 
         @user.name = "Name"
-        @attacher.activerecord_atomic_persist(reload: -> (&block) { block.call(@user) })
+        @attacher.atomic_persist(reload: -> (&block) { block.call(@user) })
 
         assert_equal "Name", @user.reload.name
       end
@@ -529,7 +502,7 @@ describe Shrine::Plugins::Activerecord do
         @user.class.update_all(avatar_data: nil) # this change will not be detected
 
         @user.name = "Name"
-        @attacher.activerecord_atomic_persist(reload: false)
+        @attacher.atomic_persist(reload: false)
 
         assert_equal "Name", @user.reload.name
       end
@@ -542,7 +515,7 @@ describe Shrine::Plugins::Activerecord do
         @user.save(validate: false)
 
         @user.name = nil
-        @attacher.activerecord_atomic_persist
+        @attacher.atomic_persist
 
         assert_nil @user.reload.name
       end
@@ -553,7 +526,7 @@ describe Shrine::Plugins::Activerecord do
         file = @attacher.attach(fakeio)
         @user.class.update_all(name: "Janko")
 
-        @attacher.activerecord_atomic_persist(nil)
+        @attacher.atomic_persist(nil)
 
         @user.reload
         @attacher.reload
@@ -569,7 +542,7 @@ describe Shrine::Plugins::Activerecord do
         @user.class.after_save { after_save_called = true }
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_persist
+        @attacher.atomic_persist
 
         assert after_save_called
       end
@@ -578,7 +551,7 @@ describe Shrine::Plugins::Activerecord do
         @attacher.attach(fakeio)
         @user.save
 
-        @attacher.activerecord_atomic_persist(persist: -> {
+        @attacher.atomic_persist(persist: -> {
           @user.name = "Janko"
           @user.save
         })
@@ -592,7 +565,7 @@ describe Shrine::Plugins::Activerecord do
         @user.save
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_persist(persist: false)
+        @attacher.atomic_persist(persist: false)
 
         assert_equal "Janko", @user.name
         assert_nil @user.reload.name
@@ -605,35 +578,31 @@ describe Shrine::Plugins::Activerecord do
         @user.class.update_all(avatar_data: file.to_json)
 
         assert_raises(Shrine::AttachmentChanged) do
-          @attacher.activerecord_atomic_persist
+          @attacher.atomic_persist
         end
 
         @user.name = "Janko"
-        @attacher.activerecord_atomic_persist(file)
+        @attacher.atomic_persist(file)
 
         assert_equal "Janko", @user.name
         assert_equal "Janko", @user.reload.name
       end
 
-      it "is aliased to #atomic_persist" do
-        file = @attacher.attach(fakeio)
-        @user.save
+      it "raises NotImplementedError for non-ActiveRecord attacher" do
+        @attacher = @shrine::Attacher.new
 
-        @user.name = "Janko"
-        @attacher.atomic_persist
-
-        assert_equal "Janko", @user.name
-        assert_equal "Janko", @user.reload.name
-        assert_equal file,    @attacher.file
+        assert_raises NotImplementedError do
+          @attacher.atomic_persist
+        end
       end
     end
 
-    describe "#activerecord_persist" do
+    describe "#persist" do
       it "persists the record" do
         file = @attacher.upload(fakeio)
         @user.avatar_data = file.to_json
 
-        @attacher.activerecord_persist
+        @attacher.persist
 
         assert_equal file.to_json, @user.reload.avatar_data
       end
@@ -645,7 +614,7 @@ describe Shrine::Plugins::Activerecord do
         file = @attacher.upload(fakeio)
         @user.avatar_data = file.to_json
 
-        @attacher.activerecord_persist
+        @attacher.persist
 
         assert_equal "Janko", @user.reload.name
       end
@@ -657,18 +626,17 @@ describe Shrine::Plugins::Activerecord do
         @user.save(validate: false)
 
         @user.name = nil
-        @attacher.activerecord_persist
+        @attacher.persist
 
         assert_nil @user.reload.name
       end
 
-      it "is aliased to #persist" do
-        file = @attacher.upload(fakeio)
-        @user.avatar_data = file.to_json
+      it "raises NotImplementedError for non-ActiveRecord attacher" do
+        @attacher = @shrine::Attacher.new
 
-        @attacher.persist
-
-        assert_equal file.to_json, @user.reload.avatar_data
+        assert_raises NotImplementedError do
+          @attacher.persist
+        end
       end
     end
   end