+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Simplify persistence interface (@janko)
+
 * `upload_options` – Keep `Shrine#_upload` private (@janko)
 
 * `infer_extension` – Keep `Shrine#basic_location` private (@janko)